+++ Changelog.md
@@ -1,6 +1,11 @@
 ## Unreleased
+<<<<<<< HEAD
 ### Updates
 - Add screenshot functionality
+=======
+## Fixed
+- Fix issue where moving the right handle past the left handle in the chart selection would break the values displayed in the selection window.
+>>>>>>> e17beb7... Swap begin and end indexes if end handle is moved past begin handle
 
 ## Version 3.0.0
 ### Changed