+++ Changelog.md
@@ -2,6 +2,16 @@
 ### Updates
 - Add screenshot functionality
 
+## Version 3.0.2
+## Fixed
+- The CSV export was exporting the wrong portion of data #123
+
+## Version 3.0.1
+## Fixed
+- Fix connecting PPK via J-Link Lite #122
+- Fix issue where moving the right handle past the left handle in the chart selection would break the values displayed in the selection window #119
+
+
 ## Version 3.0.0
 ### Changed
 - Complete rewrite of the application. This version supports the old power profiler kit, in addition to the newly released PPK2 hardware.